backend/rag_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
sentence-transformers==2.5.1
huggingface_hub==0.20.3
numpy==1.24.3
faiss-cpu==1.7.4
openai==1.51.0
httpx==0.27.2
httpcore==1.0.9
python-magic==0.4.27
openpyxl==3.1.2
requests==2.31.0
urllib3==1.26.18
pika==1.3.2
python-dotenv==1.0.0 